Change web_desktop logic to make an API call for token/sig, then build URL
